+++ pkg/backends/null/null.go
@@ -40,6 +40,6 @@ func (client client) SendEvent(ctx context.Context, e *gostatsd.Event) error {
 }
 
 // BackendName returns the name of the backend.
-func (client client) BackendName() string {
+func (client client) Name() string {
 	return BackendName
 }
