+++ pkg/backends/statsdaemon/statsdaemon_test.go
